Projeto concluído

+++ api/v3/endpoints/artigo.py
@@ -6,7 +6,7 @@ from sqlalchemy.future import select
 
 from models.artigo_model import ArtigoModel
 from models.usuario_Model import UsuarioModel
-from schemas.artigo_schema import ArtigoSchema
+from schemas.artigo_schema import ArtigoSchema,ArtigoSchemaUpdate
 from core.deps import get_session,getCurrentUser
 
 router = APIRouter()
@@ -18,14 +18,14 @@ async def postArtigo(artigo: ArtigoSchema, UsuarioLogado: UsuarioModel = Depends
         titulo=artigo.titulo,
         descricao=artigo.descricao,
         url_fonte=artigo.url_fonte,
-        usuario_id=UsuarioLogado.id,
+        usuario_id=UsuarioLogado.id, #independente do que estou mandando no body ele ta pegando o usuário autenticado pra enviar no post, por isso todos os artigos estão caindo com o mesmo ususario criado.
     )
     db.add(novoArtigo)
     await db.commit()
     return novoArtigo
 
 #get artigos
-@router.get('/',response_model=ArtigoSchema)
+@router.get('/',response_model=List[ArtigoSchema])
 async def getArtigos(db:AsyncSession = Depends(get_session)):
     async with db as session:
         querie = select(ArtigoModel)
@@ -33,10 +33,9 @@ async def getArtigos(db:AsyncSession = Depends(get_session)):
         artigos: List[ArtigoModel] = result.scalars().unique().all()
         
         if artigos:
-            
             return artigos
         else:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Artigos não encotrados.')
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Artigos não encontrados.')
 
 #get artigo
 @router.get('/{artigoID}',response_model=ArtigoSchema,status_code=status.HTTP_200_OK)
@@ -44,7 +43,7 @@ async def getArtigo(artigoID:int, db:AsyncSession = Depends(get_session)):
     async with db as session:
         querie = select(ArtigoModel).filter(ArtigoModel.id == artigoID)
         result = await session.execute(querie)
-        artigo = ArtigoModel = result.scalars().unique().one_or_none()
+        artigo:ArtigoModel = result.scalars().unique().one_or_none()
         
         if artigo:
             return artigo
@@ -53,40 +52,46 @@ async def getArtigo(artigoID:int, db:AsyncSession = Depends(get_session)):
 
 #put artigo
 @router.put('/{artigoID}',response_model=ArtigoSchema,status_code=status.HTTP_202_ACCEPTED)
-async def putArtigo(artigoID:int,artigo:ArtigoSchema,db:AsyncSession = Depends(get_session),usuarioLogado:UsuarioModel=Depends(getCurrentUser)):
+async def putArtigo(artigoID:int,artigo:ArtigoSchemaUpdate,db:AsyncSession = Depends(get_session),usuarioLogado:UsuarioModel=Depends(getCurrentUser)):
     async with db as session:
         querie = select(ArtigoModel).filter(ArtigoModel.id==artigoID)
         result = await session.execute(querie)
         artigoUP:ArtigoModel = result.scalars().unique().one_or_none()
         
         if artigoUP:
-            if artigo.titulo:
-                artigoUP.titulo = artigo.titulo
-            if artigo.descricao:
-                artigoUP.descricao = artigo.descricao
-            if artigo.url_fonte:
-                artigoUP.url_fonte = artigo.url_fonte
-            if usuarioLogado.id != artigoUP.usuario_id:
-                artigoUP.usuario_id = usuarioLogado.id
-            await session.commit()
-            return artigoUP
+            if artigoUP.usuario_id == usuarioLogado.id:
+                if artigo.titulo:
+                    artigoUP.titulo = artigo.titulo
+                if artigo.descricao:
+                    artigoUP.descricao = artigo.descricao
+                if artigo.url_fonte:
+                    artigoUP.url_fonte = artigo.url_fonte
+                if usuarioLogado.id != artigoUP.usuario_id:
+                    artigoUP.usuario_id = usuarioLogado.id
+                await session.commit()
+                return artigoUP
+            else:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='Artigo ainda não liberado para reviews públicas.')
         else:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                                detail='Artigo não encontrado')
+                                detail='Artigo não encontrado.')
             
 #delete artigo
 @router.delete ('/{artigoID}',status_code=status.HTTP_204_NO_CONTENT)
 async def deleteArtigo(artigoID:int,db:AsyncSession=Depends(get_session),usuarioLogado:UsuarioModel = Depends(getCurrentUser)):
     async with db as session:
-        querie = select(ArtigoModel).filter(ArtigoModel.id==artigoID).filter(ArtigoModel.usuario_id == usuarioLogado.id)
+        querie = select(ArtigoModel).filter(ArtigoModel.id==artigoID)#.filter(ArtigoModel.usuario_id == usuarioLogado.id)
         result = await session.execute(querie)
         artigoDel:ArtigoModel = result.scalars().unique().one_or_none()
         
         if artigoDel:
-            await session.delete(artigoDel)
-            await session.commit()
+            if usuarioLogado.id == artigoDel.usuario_id:
+                await session.delete(artigoDel)
+                await session.commit()
             
-            return Response(status_code=status.HTTP_204_NO_CONTENT)
+                return Response(status_code=status.HTTP_204_NO_CONTENT)
+            else:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='Apenas o criador do artigo pode excluir o trabalho.')
         else:
             raise HTTPException(detail='Artigo não encontrado',
                                 status_code=status.HTTP_404_NOT_FOUND)
